refactor(torghut): normalize tca modules
