+++ services/torghut/app/trading/tca_modules/__init__.py
@@ -1,63 +1,51 @@
-# pyright: reportMissingImports=false, reportUnknownVariableType=false, reportUnknownMemberType=false, reportUnknownArgumentType=false, reportUnknownParameterType=false, reportUnknownLambdaType=false, reportUnusedImport=false, reportUnusedClass=false, reportUnusedFunction=false, reportUnusedVariable=false, reportUndefinedVariable=false, reportUnsupportedDunderAll=false, reportAttributeAccessIssue=false, reportUntypedBaseClass=false, reportGeneralTypeIssues=false, reportInvalidTypeForm=false, reportReturnType=false, reportOptionalMemberAccess=false, reportArgumentType=false, reportCallIssue=false, reportPrivateUsage=false, reportUnnecessaryComparison=false, reportMissingTypeStubs=false, reportUnnecessaryCast=false
 from __future__ import annotations
 
-from importlib import import_module as __compat_import_module__
-import logging as __compat_logging__
-import sys as __compat_sys__
-import types as __compat_types__
+import logging
 
-__compat_part_modules__: list[__compat_types__.ModuleType] = []
-
-
-class __CompatModule__(__compat_types__.ModuleType):
-    def __setattr__(self, name: str, value: object) -> None:
-        super().__setattr__(name, value)
-        for module in __compat_part_modules__:
-            module.__dict__[name] = value
-
-
-def __compat_export__(module: __compat_types__.ModuleType) -> None:
-    for name, value in module.__dict__.items():
-        if name.startswith("__"):
-            continue
-        globals()[name] = value
-
-
-__compat_module__ = __compat_import_module__(f"{__name__}.part_01_statements_27")
-__compat_part_modules__.append(__compat_module__)
-__compat_export__(__compat_module__)
-for __compat_loaded_module__ in __compat_part_modules__:
-    __compat_loaded_module__.__dict__.update(
-        {name: value for name, value in globals().items() if not name.startswith("__")}
-    )
-
-__compat_module__ = __compat_import_module__(
-    f"{__name__}.part_02_observed_broker_order_policy_hash"
+from ...config import settings
+from .adaptive_policy import (
+    ADAPTIVE_DEGRADE_FALLBACK_BPS,
+    ADAPTIVE_EXECUTION_SLOWDOWN,
+    ADAPTIVE_EXECUTION_SPEEDUP,
+    ADAPTIVE_LOOKBACK_WINDOW,
+    ADAPTIVE_MAX_SHORTFALL,
+    ADAPTIVE_MAX_SLIPPAGE_BPS,
+    ADAPTIVE_MIN_EXPECTED_SHORTFALL_COVERAGE,
+    ADAPTIVE_MIN_SAMPLE_SIZE,
+    ADAPTIVE_PARTICIPATION_RELAX,
+    ADAPTIVE_PARTICIPATION_TIGHTEN,
+    ADAPTIVE_TARGET_SLIPPAGE_BPS,
+    AdaptiveExecutionPolicyDecision,
+    derive_adaptive_execution_policy,
 )
-__compat_part_modules__.append(__compat_module__)
-__compat_export__(__compat_module__)
-for __compat_loaded_module__ in __compat_part_modules__:
-    __compat_loaded_module__.__dict__.update(
-        {name: value for name, value in globals().items() if not name.startswith("__")}
-    )
-
-__compat_module__ = __compat_import_module__(
-    f"{__name__}.part_03_execution_lineage_summary"
+from .execution_tca_metrics import (
+    EXECUTION_TCA_COST_LINEAGE_SCHEMA_VERSION,
+    POST_COST_PNL_BASIS,
+    upsert_execution_tca_metric,
 )
-__compat_part_modules__.append(__compat_module__)
-__compat_export__(__compat_module__)
-for __compat_loaded_module__ in __compat_part_modules__:
-    __compat_loaded_module__.__dict__.update(
-        {name: value for name, value in globals().items() if not name.startswith("__")}
-    )
+from .lineage_read_model import build_tca_gate_inputs, refresh_execution_tca_metrics
 
-__compat_sys__.modules[__name__].__class__ = __CompatModule__
-logger = __compat_logging__.getLogger(__name__.removesuffix("_modules"))
-for __compat_loaded_module__ in globals().get("__compat_part_modules__", ()):
-    __compat_loaded_module__.__dict__["logger"] = logger
+logger = logging.getLogger("app.trading.tca")
+
 __all__ = [
-    name
-    for name in globals()
-    if not name.startswith("__") and not name.startswith("_CompatModule")
+    "ADAPTIVE_DEGRADE_FALLBACK_BPS",
+    "ADAPTIVE_EXECUTION_SLOWDOWN",
+    "ADAPTIVE_EXECUTION_SPEEDUP",
+    "ADAPTIVE_LOOKBACK_WINDOW",
+    "ADAPTIVE_MAX_SHORTFALL",
+    "ADAPTIVE_MAX_SLIPPAGE_BPS",
+    "ADAPTIVE_MIN_EXPECTED_SHORTFALL_COVERAGE",
+    "ADAPTIVE_MIN_SAMPLE_SIZE",
+    "ADAPTIVE_PARTICIPATION_RELAX",
+    "ADAPTIVE_PARTICIPATION_TIGHTEN",
+    "ADAPTIVE_TARGET_SLIPPAGE_BPS",
+    "AdaptiveExecutionPolicyDecision",
+    "EXECUTION_TCA_COST_LINEAGE_SCHEMA_VERSION",
+    "POST_COST_PNL_BASIS",
+    "build_tca_gate_inputs",
+    "derive_adaptive_execution_policy",
+    "logger",
+    "refresh_execution_tca_metrics",
+    "settings",
+    "upsert_execution_tca_metric",
 ]
-del __compat_module__
