+++ services/torghut/tests/tca_adaptive_policy/test_adaptive_policy_derivation.py
@@ -378,7 +378,7 @@ class TestAdaptiveExecutionPolicyDerivationPart1(
                 patch.object(tca_module.settings, "tigerbeetle_required", False),
                 patch(
                     (
-                        "app.trading.tca_modules.part_01_statements_27"
+                        "app.trading.tca_modules.lineage_read_model"
                         ".TigerBeetleLedgerJournal.journal_execution"
                     ),
                     side_effect=RuntimeError("journal failed"),
@@ -402,7 +402,7 @@ class TestAdaptiveExecutionPolicyDerivationPart1(
                 patch.object(tca_module.settings, "tigerbeetle_required", True),
                 patch(
                     (
-                        "app.trading.tca_modules.part_01_statements_27"
+                        "app.trading.tca_modules.lineage_read_model"
                         ".TigerBeetleLedgerJournal.journal_execution"
                     ),
                     side_effect=RuntimeError("journal failed"),
